+++ jsonargparse/core.py
@@ -261,6 +261,9 @@ class ArgumentParser(_ActionsContainer, argparse.ArgumentParser, LoggerProperty)
 
 
     def _parse_optional(self, arg_string):
+        subclass_arg = ActionTypeHint.parse_subclass_arg(self, arg_string)
+        if subclass_arg:
+            return subclass_arg
         if arg_string == self._print_config:
             arg_string += '='
         return super()._parse_optional(arg_string)