+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ v3.??.? (2021-??-??)
 Fixed
 ^^^^^
 - `--help` broken when default_config_files fail to parse #60.
+- `--*.class_path` and `--*.init_args.*` arguments not being parsed.
 
 
 v3.15.0 (2021-06-22)